Main.rb requested change
main.rb is only to start the code and ask the user for the initial choice of options
